Did DisplayHomeInteractor incorrectly

+++ WiselyWearDemo/main/java/use_case/display_home/DisplayHomeOutputData.java
@@ -17,7 +17,7 @@ public class DisplayHomeOutputData {
         return location;
     }
 
-    public int getTemperature() {
+    public double getTemperature() {
         return temperature;
     }
